pyproject.toml
[tool.poetry]
name = "python development is a coursework"
version = "0.1.0"
description = ""
authors = ["Ershov Anton <ant_e@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
mypy = "^1.10.0"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

exclude = 'env'
warn_return_any = true
no_implicit_optional = true
disallow_untyped_defs = true

[tool.isort]
line_length = 119

[tool.black]
line-length = 119

[tool.mypy]
warn_return_any = true
no_implicit_optional = true
disallow_untyped_defs = true
exclude = 'env'